+++ src/ui/common/src/components/cards/GettingStartedTutorial.tsx
@@ -76,7 +76,6 @@ const GettingStartedTutorial: React.FC<GettingStartedTutorialProps> = ({
     <Box
       sx={{
         width: '100%',
-        height: '100%',
         display: 'flex',
         justifyContent: 'center',
         alignItems: 'center',
