+++ src/ui/common/src/components/layouts/menuSidebar.tsx
@@ -26,7 +26,10 @@ import {
 import NotificationsPopover from '../notifications/NotificationsPopover';
 import styles from './menu-sidebar-styles.module.css';
 
-export const MenuSidebarWidthNumber = 200;
+// Left padding = 13px
+// Right padding = 13px
+// Content size = 50px
+export const MenuSidebarWidthNumber = 76;
 export const MenuSidebarWidth = `${MenuSidebarWidthNumber}px`;
 
 export type SidebarButtonProps = {
@@ -109,24 +112,13 @@ const SidebarButton: React.FC<SidebarButtonProps> = ({
 /**
  * The `MenuSidebar` is the core sidebar that we include throughout our UI. It
  * is pinned on the left-hand side of every page in our UI, and it includes
- * information about the user that logged in and quick links to core
- * abstractions in our system (workflows, integrations, etc).
+ * quick links to core abstractions in our system (workflows, integrations, etc).
  */
 const MenuSidebar: React.FC<{ user: UserProfile }> = ({ user }) => {
-  const [anchorEl, setAnchorEl] = useState(null);
-  const [userPopoverAnchorEl, setUserPopoverAnchorEl] = useState(null);
-  const [currentPage, setCurrentPage] = useState(undefined);
   const dispatch: AppDispatch = useDispatch();
+  const [currentPage, setCurrentPage] = useState(undefined);
   const location = useLocation();
 
-  const numUnreadNotifications = useSelector(
-    (state: RootState) =>
-      state.notificationsReducer.notifications.filter(
-        (notification) =>
-          notification.level !== NotificationLogLevel.Success &&
-          notification.status === NotificationStatus.Unread
-      ).length
-  );
 
   useEffect(() => {
     setCurrentPage(location.pathname);
@@ -135,43 +127,6 @@ const MenuSidebar: React.FC<{ user: UserProfile }> = ({ user }) => {
       dispatch(handleFetchNotifications({ user }));
     }
   }, []);
-
-  const handleClick = (event: React.MouseEvent<HTMLDivElement>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleUserPopoverClick = (event: React.MouseEvent) => {
-    setUserPopoverAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  const handleCloseUserPopover = () => {
-    setUserPopoverAnchorEl(null);
-  };
-
-  const open = Boolean(anchorEl);
-  const notificationsPopoverId = open ? 'simple-popover' : undefined;
-
-  const userPopoverOpen = Boolean(userPopoverAnchorEl);
-  const userPopoverId = userPopoverOpen ? 'user-popover' : undefined;
-
-  const avatar = user.picture ? (
-    <Avatar
-      className={styles['user-avatar']}
-      sx={{ width: '24px', height: '24px' }}
-      src={user.picture}
-    />
-  ) : (
-    <Avatar
-      className={styles['user-avatar']}
-      sx={{ width: '24px', height: '24px' }}
-    >
-      {user.name !== 'aqueduct user' ? user.name : null}
-    </Avatar>
-  );
 
   const pathPrefix = getPathPrefix();
   const sidebarContent = (
@@ -183,90 +138,16 @@ const MenuSidebar: React.FC<{ user: UserProfile }> = ({ user }) => {
           component={RouterLink as any}
         >
           <img
-            style={{ maxWidth: '250px', width: '100%', marginLeft: '4px' }}
-            src="https://aqueduct-public-assets-bucket.s3.us-east-2.amazonaws.com/webapp/logos/aqueduct-logo-horizontal-light-transparent-resized.png"
+            src={
+              'https://aqueduct-public-assets-bucket.s3.us-east-2.amazonaws.com/webapp/logos/aqueduct_logo_color_on_white.png'
+            }
+            width="50px"
+            height="50px"
           />
         </Link>
-
-        {/* popover target */}
-        <Button
-          sx={{
-            ...BUTTON_STYLE_OVERRIDE,
-            mt: 2,
-            mb: 1,
-            height: '40px',
-            backgroundColor: 'gray.100',
-            color: 'darkGray',
-            '&:hover': {
-              backgroundColor: 'gray.300',
-            },
-            alignItems: 'center',
-          }}
-          onClick={handleUserPopoverClick}
-          disableRipple
-        >
-          {avatar}
-          <Box
-            sx={{
-              textOverflow: 'clip',
-              whiteSpace: 'nowrap',
-              display: 'block',
-              overflow: 'hidden',
-              width: '130px',
-              maxWidth: '130px',
-              fontSize: '16px',
-              ml: 1,
-            }}
-          >
-            {user.name === 'aqueduct user' ? 'Aqueduct' : user.name}
-          </Box>
-        </Button>
-        {/* end popover target */}
-
-        <Menu
-          id={userPopoverId}
-          anchorEl={userPopoverAnchorEl}
-          onClose={handleCloseUserPopover}
-          open={userPopoverOpen}
-        >
-          <Link
-            to={`${getPathPrefix()}/account`}
-            underline="none"
-            sx={{ color: 'blue.800' }}
-            component={RouterLink as any}
-          >
-            <MenuItem sx={{ width: '190px' }} disableRipple>
-              <Box sx={{ fontSize: '20px', mr: 1 }}>
-                <FontAwesomeIcon icon={faCircleUser} />
-              </Box>
-              Account
-            </MenuItem>
-          </Link>
-        </Menu>
       </Box>
 
-      <Box className={styles['menu-sidebar-links']}>
-        <Box className={styles['menu-sidebar-links-wrapper']}>
-          <Box className={styles['menu-sidebar-link']} onClick={handleClick}>
-            <SidebarButton
-              icon={
-                <FontAwesomeIcon
-                  className={styles['menu-sidebar-icon']}
-                  icon={faBell}
-                />
-              }
-              text="Notifications"
-              numUpdates={numUnreadNotifications}
-            />
-          </Box>
-
-          <NotificationsPopover
-            user={user}
-            id={notificationsPopoverId}
-            anchorEl={anchorEl}
-            handleClose={handleClose}
-            open={open}
-          />
+      {/* <Box className={styles['menu-sidebar-links']}>
 
           <Link
             to={`${getPathPrefix()}/workflows`}
@@ -339,7 +220,7 @@ const MenuSidebar: React.FC<{ user: UserProfile }> = ({ user }) => {
             </Link>
           </Box>
         </Box>
-      </Box>
+      </Box> */}
     </>
   );
